% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/business_suffix.R
\name{business_suffix}
\alias{business_suffix}
\title{Business name suffix normalization}
\usage{
business_suffix(vect)
}
\arguments{
\item{vect}{Character vector of business names.}
}
\value{
Character vector of business names, with business name suffixes
  edited to be normalized.
}
\description{
Function that attempts to merge common business name suffixes within a
character string.
}
\details{
NOTE: This function also edits all instances of the word "and" to
  be "&". Here is a complete list of the edits this function will make:
  \itemize{
  \item "incorporated" and "incorporate" become "inc".
  \item "corporation" and  "corporations" become "corp".
  \item "company", "companys" and "companies" become "co".
  \item "limited liability company" becomes "llc".
  \item "limited" becomes "ltd".
  \item "division" and  "divisions" become "div".
  \item "enterprises" and "enterprise" become "ent".
  \item "limited partnership" becomes "lp".
  \item "and" becomes "&".
  }
}
\examples{
\dontrun{
business_suffix(c("Acme Inc", "Acme incorporated", "acme company"))
[1] "Acme Inc" "Acme inc" "acme co"
}
}
