% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{equality}
\alias{equality}
\title{Compare a char string and char vector}
\usage{
equality(lookupvect, charstring)
}
\arguments{
\item{lookupvect}{Character vector.}

\item{charstring}{Character string.}
}
\description{
Takes a char vector and char string as input, compares the string to every
element of the char vector, output is a logical vector with length equal to
the input char vector. This is replacing the R code "vector == string".
}
