% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fingerprint_KC.R
\name{get_fingerprint_KC}
\alias{get_fingerprint_KC}
\title{Get key collision fingerprints}
\usage{
get_fingerprint_KC(vect, bus_suffix = TRUE, ignore_strings = NULL)
}
\arguments{
\item{vect}{Character vector.}

\item{bus_suffix}{Logical indicating whether the merging of records should
be insensitive to common business suffixes (TRUE) or not (FALSE). If
input \code{vect} a vector of business names it's recommended to set this
to TRUE. Default value is TRUE.}

\item{ignore_strings}{Character vector, these strings will be ignored during
the merging of values within \code{vect}. Default value is NULL.}
}
\value{
Key collision value of the input string.
}
\description{
Given a character vector as input, get the key collision fingerprint for
each element. For more details on key collision, see
\url{https://github.com/OpenRefine/OpenRefine/wiki/Clustering-In-Depth}.
}
\examples{
\dontrun{
get_fingerprint_KC("Tom's Sports Equipment, Inc.")
[1] "equipment sports toms"
}
}
