% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fingerprint_ngram.R
\name{get_fingerprint_ngram}
\alias{get_fingerprint_ngram}
\title{Get ngram fingerprints}
\usage{
get_fingerprint_ngram(vect, numgram, bus_suffix)
}
\arguments{
\item{vect}{Character vector.}

\item{numgram}{Numeric value indicating the number of characters that
will occupy each token (default value is 2).}

\item{bus_suffix}{Logical indicating whether the merging of records should
be insensitive to common business suffixes (TRUE) or not (FALSE). If
working with a vector of business names it's recommended to set this to
TRUE. Default value is TRUE.}
}
\value{
Ngram values of the input vector.
}
\description{
Given a character vector as input, get the ngram fingerprint value for each
element of the input.
For more details on ngram fingerprinting, see
\url{https://github.com/OpenRefine/OpenRefine/wiki/Clustering-In-Depth}.
}
\examples{
\dontrun{
get_fingerprint_ngram("Tom's Sports Equipment, Inc.", numgram = 1)
[1] "eimnopqrstu"
get_fingerprint_ngram("Tom's Sports Equipment, Inc.", numgram = 2)
[1] "eneqipmemsntomorpmpoqurtsespsstotsui"
}
}
