% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_collision_merge.R
\name{key_collision_merge}
\alias{key_collision_merge}
\title{Value merging based on Key Collision}
\usage{
key_collision_merge(vect, bus_suffix = TRUE)
}
\arguments{
\item{vect}{Character vector of items for which you want similar values
merged.}

\item{bus_suffix}{Logical indicating whether the merging of records should
be insensitive to common business suffixes (TRUE) or not (FALSE). If
working with a vector of business names it's recommended to set this to
TRUE. Default value is TRUE.}
}
\value{
Character vector with similar values merged.
}
\description{
This function takes a character vector and makes edits and merges values
that are approximately equivalent yet not identical. It clusters values
based on the key collision method, described here
\url{https://github.com/OpenRefine/OpenRefine/wiki/Clustering-In-Depth}.
}
\examples{
\dontrun{
x <- c("Acme Pizza, Inc.",
       "ACME PIZZA COMPANY",
       "pizza, acme llc",
       "Acme Pizza, Inc.")
key_collision_merge(vect = x, bus_suffix = TRUE)
[1] "Acme Pizza, Inc." "Acme Pizza, Inc." "Acme Pizza, Inc." "Acme Pizza, Inc."
}
}
