% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_lens}
\alias{list_lens}
\title{Get Length of Each List Element}
\usage{
list_lens(x)
}
\arguments{
\item{x}{List}
}
\value{
Integer vector, indicating the length of each input list element.
}
\description{
R wrapper for C function `C_list_lens()` (`C_list_lens()` calls
stringdist C function `R_lengths()`).
}
\examples{
test_list <- list(c(1, 2, 3), c("cats", "dogs"))
list_lens(test_list)
}
