% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_clusters}
\alias{merge_clusters}
\title{Merge clusters of similar values.}
\usage{
merge_clusters(clusters, keys, vect, keyssub, vectsub)
}
\arguments{
\item{clusters}{character vector}

\item{keys}{character vector}

\item{vect}{character vector}

\item{keyssub}{character vector}

\item{vectsub}{character vector}
}
\description{
Function that performs all merges related to input value clusters.
}
