% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_clusters_dict}
\alias{merge_clusters_dict}
\title{Function that performs all merges related to input value clusters. Can
perform merging using values from a dictionary vector.}
\usage{
merge_clusters_dict(clusters, keys_vect, vect, keys_vect_sub, vect_sub,
  keys_dict, dict)
}
\arguments{
\item{clusters}{character vector}

\item{keys_vect}{character vector}

\item{vect}{character vector}

\item{keys_vect_sub}{character vector}

\item{vect_sub}{character vector}

\item{keys_dict}{character vector}

\item{dict}{character vector}
}
\description{
Function that performs all merges related to input value clusters. Can
perform merging using values from a dictionary vector.
}
