% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_ngram_clusters_string}
\alias{merge_ngram_clusters_string}
\title{Merge values related to a cluster of length 1}
\usage{
merge_ngram_clusters_string(cluster, n_gram_keys, univect, vect)
}
\arguments{
\item{cluster}{character string}

\item{n_gram_keys}{character vector}

\item{univect}{character vector}

\item{vect}{character vector}
}
\description{
Function that performs all merges related to a single cluster, given that
the input cluster is a string (char vector of length 1). Values within
arg vect that are related to the cluster string are edited/merged, return
is the edited char vector "vect".
}
