% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_ngram_clusters_vector}
\alias{merge_ngram_clusters_vector}
\title{Merge values related to a cluster of length > 1}
\usage{
merge_ngram_clusters_vector(cluster, n_gram_keys, univect, vect)
}
\arguments{
\item{cluster}{character vector}

\item{n_gram_keys}{character vector}

\item{univect}{character vector}

\item{vect}{character vector}
}
\description{
Function that performs all merges related to a single cluster, given that
the input cluster is a char vector with length greater than one. Values
within arg vect that are related to the cluster are edited/merged, returnm
is the edited char vector "vect".
}
