% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{most_freq}
\alias{most_freq}
\title{Find the string that's most freqent in a vector.}
\usage{
most_freq(clust, keyssub, vectsub)
}
\arguments{
\item{clust}{character string}

\item{keyssub}{character vector}

\item{vectsub}{character vector}
}
\description{
Given a cluster suitable for merging, find all instances within vectsub
that coorspond with that cluster and get the most freqently occuring
value (this is basically using freq to choose the value from the original
data that will be used as the template for all elements of a single cluster).
eg, if we have a cluster value that coorespond with these values from the
original data: c("Bob's Pizza", "bobs pizza", "Bob's Pizza"), then all those
values would be edited to be "Bob's Pizza", since its the most frequent.
}
