% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Rcpp_str_dist_all_int}
\alias{Rcpp_str_dist_all_int}
\title{Check to see if all elements of a list are integer vectors}
\usage{
Rcpp_str_dist_all_int(x)
}
\arguments{
\item{x}{List}
}
\value{
Logical
}
\description{
R wrapper for C function `C_test_all_int()` (`C_test_all_int()`
calls stringdist C function `R_all_int()`)
}
\examples{
test_list <- list(c(1, 2, 3), c("cats", "dogs"))
str_dist_all_int(test_list)
}
