% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Rcpp_str_dist_lengths}
\alias{Rcpp_str_dist_lengths}
\title{Get Length of Each List Element}
\usage{
Rcpp_str_dist_lengths(x)
}
\arguments{
\item{x}{List}
}
\value{
Integer vector, indicating the length of each input list element.
}
\description{
R wrapper for C function `C_test_lengths()` (`C_test_lengths()`
calls stringdist C function `R_lengths()`).
}
\examples{
test_list <- list(c(1, 2, 3), c("cats", "dogs"))
str_dist_lengths(test_list)
}
