% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_clusters}
\alias{merge_clusters}
\title{Merge clusters of similar values.}
\usage{
merge_clusters(clusters, keys_vect, vect, keys_vect_sub, vect_sub)
}
\arguments{
\item{clusters}{character vector}

\item{keys_vect}{character vector}

\item{vect}{character vector}

\item{keys_vect_sub}{character vector}

\item{vect_sub}{character vector}
}
\description{
Function that performs all merges related to input value clusters.
}
